handel a product card
